Fix lỗi admin là duy nhất

+++ resources/views/backend/admin/Role/role.blade.php
@@ -25,22 +25,41 @@
             <th width="280px">Action</th>
         </tr>
         @foreach ($roles as $key => $role)
-            <tr>
-                <td>{{ ++$i }}</td>
-                <td>{{ $role->display_name }}</td>
-                <td>{{ $role->description }}</td>
-                <td>
-                    <a class="btn btn-info" href="{{ route('roles.show',$role->id) }}">Show</a>
-                    @permission(('role-edit'))
-                    <a class="btn btn-primary" href="{{ route('roles.edit',$role->id) }}">Edit</a>
-                    @endpermission
-                    @permission(('role-delete'))
-                    {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
-                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                    {!! Form::close() !!}
-                    @endpermission
-                </td>
-            </tr>
+            @if(Auth::user()->hasRole('admin'))
+                <tr>
+                    <td>{{ ++$i }}</td>
+                    <td>{{ $role->display_name }}</td>
+                    <td>{{ $role->description }}</td>
+                    <td>
+                        <a class="btn btn-info" href="{{ route('roles.show',$role->id) }}">Show</a>
+                        @permission(('role-edit'))
+                        <a class="btn btn-primary" href="{{ route('roles.edit',$role->id) }}">Edit</a>
+                        @endpermission
+                        @permission(('role-delete'))
+                        {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
+                        {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                        {!! Form::close() !!}
+                        @endpermission
+                    </td>
+                </tr>
+            @elseif(!($role->name=='admin'))
+                <tr>
+                    <td>{{ ++$i }}</td>
+                    <td>{{ $role->display_name }}</td>
+                    <td>{{ $role->description }}</td>
+                    <td>
+                        <a class="btn btn-info" href="{{ route('roles.show',$role->id) }}">Show</a>
+                        @permission(('role-edit'))
+                        <a class="btn btn-primary" href="{{ route('roles.edit',$role->id) }}">Edit</a>
+                        @endpermission
+                        @permission(('role-delete'))
+                        {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'style'=>'display:inline']) !!}
+                        {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                        {!! Form::close() !!}
+                        @endpermission
+                    </td>
+                </tr>
+            @endif
         @endforeach
     </table>
     {!! $roles->render() !!}
