+++ resources/views/backend/admin/User/user.blade.php
@@ -24,25 +24,55 @@
             <th width="280px">Action</th>
         </tr>
         @foreach ($data as $key => $user)
-            <tr>
-                <td>{{ ++$i }}</td>
-                <td>{{ $user->name }}</td>
-                <td>{{ $user->email }}</td>
-                <td>
-                    @if(!empty($user->roles))
-                        @foreach($user->roles as $v)
-                            <label class="label label-success">{{ $v->display_name }}</label>
-                        @endforeach
-                    @endif
-                </td>
-                <td>
-                    <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
-                    <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
-                    {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
-                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                    {!! Form::close() !!}
-                </td>
-            </tr>
+            @if(Auth::user()->hasRole('admin'))
+                <tr>
+                    <td>{{ ++$i }}</td>
+                    <td>{{ $user->name }}</td>
+                    <td>{{ $user->email }}</td>
+                    <td>
+                        @if(!empty($user->roles))
+                            @foreach($user->roles as $v)
+                                <label class="label label-success">{{ $v->display_name }}</label>
+                            @endforeach
+                        @endif
+                    </td>
+                    <td>
+                        <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+                        @permission(('user-edit'))
+                        <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+                        @endpermission
+                        @permission(('user-delete'))
+                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                        {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                        {!! Form::close() !!}
+                        @endpermission
+                    </td>
+                </tr>
+            @elseif(!$user->hasRole('admin'))
+                <tr>
+                    <td>{{ ++$i }}</td>
+                    <td>{{ $user->name }}</td>
+                    <td>{{ $user->email }}</td>
+                    <td>
+                        @if(!empty($user->roles))
+                            @foreach($user->roles as $v)
+                                <label class="label label-success">{{ $v->display_name }}</label>
+                            @endforeach
+                        @endif
+                    </td>
+                    <td>
+                        <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+                        @permission(('user-edit'))
+                        <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+                        @endpermission
+                        @permission(('user-delete'))
+                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                        {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                        {!! Form::close() !!}
+                        @endpermission
+                    </td>
+                </tr>
+            @endif
         @endforeach
     </table>
     {!! $data->links() !!}
